pyproject.toml
[project]
name = "authenticated-key-exchange"
version = "0.1.0"
description = "Authenticated Key Exchange P79 Lab 02"
requires-python = ">=3.13"
license = { file = "LICENSE" }
authors = [
    { name = "Christoph Schnabl", email = "cs2280@cam.ac.uk" },
]
readme = { file = "README.md", content-type = "text/markdown" }
dependencies = [
    "pynacl>=1.5.0",
    "pydantic>=2.10.0",
]

[project.urls]
homepage = "https://github.com/chrisschnabl/authenticated-key-exchange#readme"
repository = "https://github.com/chrisschnabl/authenticated-key-exchange"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/certificates", "src/models", "src/network", "src/sigma"]

[dependency-groups]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "mypy>=1.14.1",
  "bandit>=1.8.0",
  "ruff>=0.8.6",
  "parameterized>=0.9.0", # For parameterized tests
  "pydantic>=2.10.0",
]

[tool.pytest.ini_options]
addopts = "-vvv"
testpaths = "tests"

[tool.ruff]
extend-exclude = [
    "__pycache__",
    "build",
    "dist",
]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
extend-select = [
    "C4",
    "D200",
    "D201",
    "D204",
    "D206",
    "D210",
    "D211",
    "D213",
    "D300",
    "D402",
    "D403",
    "D404",
    "D419",
    "E",
    "F",
    "G010",
    "I",
    "INP001",
    "N805",
    "PERF101",
    "PERF102",
    "PERF401",
    "PERF402",
    "PGH004",
    "PGH005",
    "PIE794",
    "PIE796",
    "PIE807",
    "PIE810",
    "RET502",
    "RET503",
    "RET504",
    "RET505",
    "RUF015",
    "RUF100",
    "S101",
    "T20",
    "UP",
    "W",
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

[tool.mypy]
files = ["src", "tests"]
strict = "true"
plugins = ["pydantic.mypy"]
